backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
openpyxl==3.1.2
pdfplumber==0.10.3
openai==1.3.7
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5

